Story 6.4: the two mobile traders (Epic 6 complete)
Add the economy's two coin sinks (FR-21, AD-17), headless-first:
- Trader: stationary killable two-kind tile-agent (WANDERER/BLACK_MARKET),
  mirrors RogueEnemy; placed from a seed-derived sub-stream (seed ^ 0x77AAD3)
  drawn last in generateFloor so pre-6.4 seeds stay byte-identical (AD-5).
- TradeSystem: single transaction authority. Scarcity spread
  priceToBuy=base*2 > base > priceToSell=base/2 (a round-trip always loses
  coin, no infinite-money loop). Value-conserving makeChange (low->high pay /
  high->low refund) circulates the Royal Gold Plaque as change and lands the
  6.3-deferred Supply.copperValue int->long.
- TradeController: safe-pause surface (AD-14), mirrors JournalController.
- Kill-lockout: attacking the Black Market routes through CombatSystem (AD-4);
  his death sets FlagStore.KEY_BLACK_MARKET_DEAD and refuses the trade forever.
- Minimal MarginScreen trade panel + tinted trader glyphs; bespoke shop UI and
  roaming deferred (6.1/6.2 UI-deferral precedent).

Code-review (3-layer) patches: spill overflow coin/items to the floor instead
of silently losing them (drop-or-leave, FR-9); gate [T] on trader adjacency so
craft-torch still works; tileFree() prevents actor stacking in placeTraders;
refuse zero-price sells; barter picks the highest-value affordable stock; drop
the dead PlayerAction.TRADE.

Suite 571 green; core + desktop compile clean. Epic 6 done (6.1-6.4).

+++ core/src/main/java/com/margins/rogue/item/Supply.java
@@ -118,14 +118,16 @@ public enum Supply {
     }
 
     /** This coin's worth in Copper (the base unit), 0 for non-currency. The chain encodes the
-     *  epics' exchange rates: Silver = 25·Copper, Gold = 10·Silver, Royal Gold Plaque = 1000·Gold. */
-    public int copperValue() {
+     *  epics' exchange rates: Silver = 25·Copper, Gold = 10·Silver, Royal Gold Plaque = 1000·Gold.
+     *  {@code long} (Story 6.4): the Royal Gold Plaque circulates once traders give change, and a
+     *  trader's bulk price math must not overflow {@code int} at large plaque counts. */
+    public long copperValue() {
         switch (this) {
-            case COPPER:            return 1;
-            case SILVER:            return 25;               // 25:1
-            case GOLD:              return 25 * 10;          // 10:1 over Silver = 250
-            case ROYAL_GOLD_PLAQUE: return 25 * 10 * 1000;   // 1000:1 over Gold = 250_000
-            default:                return 0;
+            case COPPER:            return 1L;
+            case SILVER:            return 25L;               // 25:1
+            case GOLD:              return 25L * 10;          // 10:1 over Silver = 250
+            case ROYAL_GOLD_PLAQUE: return 25L * 10 * 1000;   // 1000:1 over Gold = 250_000
+            default:                return 0L;
         }
     }
